Added AccountHandler.cpp in makefile branch -- Improving codes by using makefile
